requirements.txt
bcrypt==4.0.1
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
idna==3.4
jmespath==1.0.1
paramiko==3.1.0
psycopg2-binary==2.9.5
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
ratelimit==2.2.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
